Refactor file type detection API

+++ src/common/AbstractSimpleIO.h
@@ -7,7 +7,7 @@
 class AbstractSimpleReader {
 public:
 	virtual ~AbstractSimpleReader() = default;
-	virtual int read(void *ptr, size_t len) = 0;
+	virtual int read(void *ptr, int64_t len) = 0;
 	virtual int64_t pos() const = 0;
 	virtual void seek(int64_t pos) = 0;
 };
@@ -15,7 +15,7 @@ public:
 class AbstractSimpleWriter {
 public:
 	virtual ~AbstractSimpleWriter() = default;
-	virtual int write(void const *ptr, size_t len) = 0;
+	virtual int write(void const *ptr, int64_t len) = 0;
 };
 
 #endif // ABSTRACTSIMPLEIO_H
